feat: LowLevelInterestManagement
A low level interest management base class that allows more advanced interest management by bypassing the built-in HashSet checks
