+++ Assets/Mirror/Runtime/NetworkClient.cs
@@ -97,7 +97,7 @@ namespace Mirror
 
         // interest management component (optional)
         // only needed for SetHostVisibility
-        public static InterestManagement aoi;
+        public static LowLevelInterestManagement aoi;
 
         // scene loading
         public static bool isLoadingScene;
