+++ Assets/Mirror/Runtime/NetworkServer.cs
@@ -43,7 +43,7 @@ namespace Mirror
 
         // interest management component (optional)
         // by default, everyone observes everyone
-        public static InterestManagement aoi;
+        public static LowLevelInterestManagement aoi;
 
         // OnConnected / OnDisconnected used to be NetworkMessages that were
         // invoked. this introduced a bug where external clients could send
@@ -202,7 +202,6 @@ namespace Mirror
             connections.Clear();
             connectionsCopy.Clear();
             handlers.Clear();
-            newObservers.Clear();
 
             // this calls spawned.Clear()
             CleanupSpawned();
@@ -1433,10 +1432,6 @@ namespace Mirror
             }
         }
 
-        // allocate newObservers helper HashSet only once
-        // internal for tests
-        internal static readonly HashSet<NetworkConnectionToClient> newObservers = new HashSet<NetworkConnectionToClient>();
-
         // rebuild observers default method (no AOI) - adds all connections
         static void RebuildObserversDefault(NetworkIdentity identity, bool initialize)
         {
@@ -1448,103 +1443,6 @@ namespace Mirror
                 if (identity.visible != Visibility.ForceHidden)
                 {
                     AddAllReadyServerConnectionsToObservers(identity);
-                }
-            }
-        }
-
-        // rebuild observers via interest management system
-        static void RebuildObserversCustom(NetworkIdentity identity, bool initialize)
-        {
-            // clear newObservers hashset before using it
-            newObservers.Clear();
-
-            // not force hidden?
-            if (identity.visible != Visibility.ForceHidden)
-            {
-                aoi.OnRebuildObservers(identity, newObservers);
-            }
-
-            // IMPORTANT: AFTER rebuilding add own player connection in any case
-            // to ensure player always sees himself no matter what.
-            // -> OnRebuildObservers might clear observers, so we need to add
-            //    the player's own connection AFTER. 100% fail safe.
-            // -> fixes https://github.com/vis2k/Mirror/issues/692 where a
-            //    player might teleport out of the ProximityChecker's cast,
-            //    losing the own connection as observer.
-            if (identity.connectionToClient != null)
-            {
-                newObservers.Add(identity.connectionToClient);
-            }
-
-            bool changed = false;
-
-            // add all newObservers that aren't in .observers yet
-            foreach (NetworkConnectionToClient conn in newObservers)
-            {
-                // only add ready connections.
-                // otherwise the player might not be in the world yet or anymore
-                if (conn != null && conn.isReady)
-                {
-                    if (initialize || !identity.observers.ContainsKey(conn.connectionId))
-                    {
-                        // new observer
-                        conn.AddToObserving(identity);
-                        // Debug.Log($"New Observer for {gameObject} {conn}");
-                        changed = true;
-                    }
-                }
-            }
-
-            // remove all old .observers that aren't in newObservers anymore
-            foreach (NetworkConnectionToClient conn in identity.observers.Values)
-            {
-                if (!newObservers.Contains(conn))
-                {
-                    // removed observer
-                    conn.RemoveFromObserving(identity, false);
-                    // Debug.Log($"Removed Observer for {gameObjec} {conn}");
-                    changed = true;
-                }
-            }
-
-            // copy new observers to observers
-            if (changed)
-            {
-                identity.observers.Clear();
-                foreach (NetworkConnectionToClient conn in newObservers)
-                {
-                    if (conn != null && conn.isReady)
-                        identity.observers.Add(conn.connectionId, conn);
-                }
-            }
-
-            // special case for host mode: we use SetHostVisibility to hide
-            // NetworkIdentities that aren't in observer range from host.
-            // this is what games like Dota/Counter-Strike do too, where a host
-            // does NOT see all players by default. they are in memory, but
-            // hidden to the host player.
-            //
-            // this code is from UNET, it's a bit strange but it works:
-            // * it hides newly connected identities in host mode
-            //   => that part was the intended behaviour
-            // * it hides ALL NetworkIdentities in host mode when the host
-            //   connects but hasn't selected a character yet
-            //   => this only works because we have no .localConnection != null
-            //      check. at this stage, localConnection is null because
-            //      StartHost starts the server first, then calls this code,
-            //      then starts the client and sets .localConnection. so we can
-            //      NOT add a null check without breaking host visibility here.
-            // * it hides ALL NetworkIdentities in server-only mode because
-            //   observers never contain the 'null' .localConnection
-            //   => that was not intended, but let's keep it as it is so we
-            //      don't break anything in host mode. it's way easier than
-            //      iterating all identities in a special function in StartHost.
-            if (initialize)
-            {
-                if (!newObservers.Contains(localConnection))
-                {
-                    if (aoi != null)
-                        aoi.SetHostVisibility(identity, false);
                 }
             }
         }
@@ -1579,7 +1477,7 @@ namespace Mirror
             // otherwise let interest management system rebuild
             else
             {
-                RebuildObserversCustom(identity, initialize);
+                aoi.OnRequestRebuild(identity, initialize);
             }
         }
